[VSC-10] Refactor test suites
Problem: We have two files for testing definitions, so we should join
into one. Moreover, we have a lot of shared things between the
definition and reference tests.

Solution: Join test cases from `src/server/definition.spec.ts` into
`src/server/test/definition.spec.ts`. Join Motoko reference test files
tests into the Motoko definitions test files. Extract common parts to
helpers. Move reference tests to `src/server/test` directory.
