+++ test/definition/A.mo
