+++ test/definition/B.mo
