+++ test/definition/circular.mo
