--- test/references/A.mo
@@ -1,9 +1,0 @@
-import B "B";
-import C "C";
-
-actor {
-    public func name() : async Nat {
-        let a = B.obj;
-        return a.meth(C.Inner.inner);
-    };
-};
