--- test/references/circular.mo
@@ -1,9 +1,0 @@
-module {
-    public class O() {
-        public let o : O = O();
-
-        public func test() : O {
-            o.o.o.o.o.o
-        }
-    };
-}
